Further pull request fixes.
Related to PRJLTVA-219.

+++ XamarinPlayer/XamarinPlayer/Services/IPlayerService.cs
@@ -24,7 +24,7 @@ namespace XamarinPlayer.Services
     {
         IObservable<PlayerState> StateChanged();
         IObservable<string> PlaybackError();
-        IObservable<double> BufferingProgress();
+        IObservable<int> BufferingProgress();
 
         TimeSpan Duration { get; }
